ai-ops-manager/backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-openai
chromadb
python-dotenv
langchain-community